refact: separate map logic into functions
