+++ src/app/equipment-tracker/ui/equipment-map/equipment-map.component.html
