+++ src/app/equipment-tracker/ui/equipment-map/equipment-map.component.scss
